refactor: #161 split Swagger API files
